requirements.txt
flask
requests
groq>=0.9.0
python-dotenv
